Ensure replication of cheops-all database
cheops-all is a meta database, it is synchronized between all sites. It
contains the sites in the cluster. When a new patch arrives in the
cheops database, we also check if the related meta database with the
same site is already known by us, and if not, create the replication,
still push-only