Transition to a DAG of operations
Each operation is its own CouchDB document, just like replies. We use that
instead of KnownState to determine the concurrency of operations.

Since each operation is its own document, it is also more efficient when
syncing (only new operations need to be synced) and specifying ancestry (when a
conflict is "resolved", the structure only contains one id)